Pass the compliance suite tests
This was tougher than I thought. The reason being that there was no good
feedback on exactly why the test might be failing.
On some occaisions it was badly implemented methods, on others it was
sneaky "undefined" variables that didn't cause exceptions, and for
others it was control flow statements not going the way I expected
and so on and so forth.

I probably would have avoided a majority of these had I written my own
initial test suite before running this full one.

Well, it is too late to complain commit and move on.